Move io core.py to crtf
